--- components/Kakao_consult.tsx
@@ -1,13 +1,0 @@
-export default function Kakao() {
-    return (
-        <div className="w-full bg-gray-100 md:bg-blue-500 md:w-6/12">
-            <div className="pl-14 pt-9">
-                <h1>도움이 필요하신가요?</h1>
-                <span className="block text-xs text-gray-400 my-3">평일 09:00 ~ 18:00 상담 가능
-                    <br />*제품교환의 경우 카톡 채널로 바로 연락 부탁드립니다.
-                </span>
-            </div>
-            <button className="w-10/12 bg-yellow-300 py-4 mx-8 rounded-lg mb-3">카톡채널 상담</button>
-        </div>
-    )
-}
