--- components/Notice.tsx
@@ -1,11 +1,0 @@
-import React, { FC, Fragment } from "react";
-
-const Notice: FC<{}> = () => {
-    return (
-        <Fragment>
-            <div>공지</div>
-        </Fragment>
-    )
-}
-
-export default Notice;
