--- components/Question.tsx
@@ -1,11 +1,0 @@
-import React, { FC, Fragment } from "react";
-
-const Question: FC<{}> = () => {
-    return (
-        <Fragment>
-            <div>자주하는 질문</div>
-        </Fragment>
-    )
-}
-
-export default Question;
