+++ pages/customer/index.tsx
@@ -1,0 +1,9 @@
+import CustomerCenter from "@/components/customer/CustomerCenter";
+
+export default function Customer() {
+    return (
+        <>
+            <CustomerCenter />
+        </>
+    )
+}
